aha, fix integer multiplication overflow

+++ lib/mergesort.hh
@@ -20,10 +20,10 @@
 
 /** @brief: Merge @a[@afirst + @astep * i] (0 <= i < @nmya), and output to @sized_output */
 template <typename C, typename F>
-void mergesort_impl(C *a, int nmya, int afirst, int astep, F &pcmp, C &sized_output) {
+void mergesort_impl(C *a, size_t nmya, size_t afirst, size_t astep, F &pcmp, C &sized_output) {
     typedef typename C::iterator iterator_type;
     xarray<iterator_type> ai;
-    for (int i = 0; i < nmya; ++i) {
+    for (size_t i = 0; i < nmya; ++i) {
         iterator_type mi = a[afirst + i * astep].begin();
         if (mi != mi.parent_end())
             ai.push_back(mi);
@@ -47,7 +47,7 @@ void mergesort_impl(C *a, int nmya, int afirst, int astep, F &pcmp, C &sized_out
 }
 
 template <typename C, typename F>
-C *mergesort(xarray<C> &a, int astep, int afirst, F &pcmp) {
+C *mergesort(xarray<C> &a, size_t astep, size_t afirst, F &pcmp) {
     size_t nmya = a.size() / astep + (size_t(afirst) < (a.size() % astep));
     size_t np = 0;
     for (size_t i = 0; i < nmya; i++)
@@ -56,7 +56,7 @@ C *mergesort(xarray<C> &a, int astep, int afirst, F &pcmp) {
     if (np == 0)
 	return out;
     mergesort_impl(a.array(), nmya, afirst, astep, pcmp, *out);
-    dprintf("mergesort: afirst %d astep %d (collections %zd : nr-kvs %zu)\n",
+    dprintf("mergesort: afirst %zd astep %zd (collections %zd : nr-kvs %zu)\n",
  	    afirst, astep, a.size(), np);
     return out;
 }
